Module Irj_utils.Test_interpreter

val check_all_tests : M_ir.Mir.program -> string -> Utils.Config.value_sort -> Utils.Config.round_ops -> (string -> bool) -> unit

check_all_tests p folder vs ro filter Executes p with all tests in folder whose name satisfy filter.

val check_one_test : M_ir.Mir.program -> string -> Utils.Config.value_sort -> Utils.Config.round_ops -> unit

Same as check_all_tests, but for one test.