Module Com.CatVar

High level representation of categories of variable. In practice, there only are two types of categories: * "calculee" (Input) variables, with several sub categories; * "saisie" (Computed) variable, that can either be base or non-base.

Such representation allows to manipulate sets of variable (for example in iterations).

type t =

| Input of Utils.StrSet.t
An input variable category with its sub categories

| Computed of {

is_base : bool;

}
A computed var with a flag marking if it is a base varible.

val all_inputs : t

A category containing all inputs.

val is_input : t -> bool

Returns true if the category in argument is an input variable category.

val is_computed : t -> bool

Returns true if the category in argument is a computed variable category.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty printer.

val compare : t -> t -> int

module Set : Utils.SetExt.T with type T.elt = t

module Map : sig ... end

type loc =

| LocComputed

| LocBase

| LocInput

val pp_loc : Stdlib.Format.formatter -> loc -> unit

module LocSet : Utils.SetExt.T with type T.elt = loc

module LocMap : Utils.MapExt.T with type T.key = loc

type data = {

id : t;
The category

id_str : string;
A unique string identifier for the category

id_int : int;
A unique int identifier

loc : loc;
Quick access to the category kind

pos : Utils.Pos.t;
Variable category declaration

attributs : Utils.Pos.t Utils.StrMap.t;
Attributes of the category

}

Data for a given category. Defined in the validator.