Module Mlang

module Bir : sig ... end

module Bir_instrumentation : sig ... end

Instrumentation of the interpreter to computer code coverage

module Bir_interface : sig ... end

Input-output management for BIR programs interpretation

module Bir_interpreter : sig ... end

Interpretation of BIR programs

module Bir_number : sig ... end

module Bir_roundops : sig ... end

module Bir_to_dgfip_c : sig ... end

This module with a single entry point generates C files from a Bir.program.

module Bir_to_java : sig ... end

module Bir_to_oir : sig ... end

Translations to BIR to OIR and vice-versa

module CharMap : sig ... end

module Cli : sig ... end

Command-line interface helpers

module Dead_code_removal : sig ... end

module DecoupledExpr : sig ... end

module Dgfip_gen_files : sig ... end

module Dgfip_m : sig ... end

module Dgfip_options : sig ... end

module Dgfip_varid : sig ... end

module Dict : sig ... end

module Driver : sig ... end

module Errors : sig ... end

module Format_bir : sig ... end

module Format_mast : sig ... end

module Format_mir : sig ... end

module Format_oir : sig ... end

module Inlining : sig ... end

module IntMap : sig ... end

module IntSet : sig ... end

module MapExt : sig ... end

module Mast : sig ... end

Abstract Syntax Tree for M

module Mast_to_mir : sig ... end

Mast to Mir translation of M programs.

module Mir : sig ... end

module Mir_dependency_graph : sig ... end

Defines the dependency graph of an M program

module Mir_interface : sig ... end

While type:Mir.program contains the AST of all rules, type:full_program also encapsulates the dependency order of the rules.

module Mir_typechecker : sig ... end

This modules defines M's static semantic. There is only one type: float. The typechecking is mostly about differentiating table from non-tables variables

module Mlexer : sig ... end

module Mparser : sig ... end

module Mpp_ast : sig ... end

module Mpp_format : sig ... end

module Mpp_frontend : sig ... end

Parses an MPP file into an MPP IR program

module Mpp_ir : sig ... end

module Mpp_ir_to_bir : sig ... end

module Mpp_lexer : sig ... end

module Mpp_parser : sig ... end

module Oir : sig ... end

module Oir_optimizations : sig ... end

module Parse_utils : sig ... end

Helpers for parsing

module Partial_evaluation : sig ... end

module Pos : sig ... end

A position in the source code is a file, as well as begin and end location of the form col:line

module Prelude : sig ... end

module SetExt : sig ... end

module SetSetExt : sig ... end

module StrMap : sig ... end

module StrSet : sig ... end

module StrSetMap : sig ... end

module StrSetSet : sig ... end

module Strings : sig ... end

module Test_interpreter : sig ... end