Module Bir_interpreter.MPFRMfInterp

type custom_float = Bir_number.MPFRNumber.t

Comes from the instantiation of the functor by a kind of floating-point value

type value =

| Number of custom_float

| Undefined

Functor-specific program values

val format_value : Stdlib.Format.formatter -> value -> unit

type var_value =

| SimpleVar of value

| TableVar of int * value array

Functor-specific variable values

val format_var_value : Stdlib.Format.formatter -> var_value -> unit

val format_var_value_with_var : Stdlib.Format.formatter -> (Bir.variable * var_value) -> unit

type ctx = {

ctx_local_vars : value Pos.marked Mir.LocalVariableMap.t;

ctx_vars : var_value Bir.VariableMap.t;

}

Interpretation context

val empty_ctx : ctx

val literal_to_value : Mir.literal -> value

val var_literal_to_var_value : var_literal -> var_value

val value_to_literal : value -> Mir.literal

val var_value_to_var_literal : var_value -> var_literal

val update_ctx_with_inputs : ctx -> Mir.literal Bir.VariableMap.t -> ctx

type run_error =

| ErrorValue of string * Pos.t

| FloatIndex of string * Pos.t

| IndexOutOfBounds of string * Pos.t

| IncorrectOutputVariable of string * Pos.t

| UnknownInputVariable of string * Pos.t

| ConditionViolated of Mir.Error.t * Bir.expression Pos.marked * (Bir.variable * var_value) list

| NanOrInf of string * Bir.expression Pos.marked

| StructuredError of string * (string option * Pos.t) list * ( unit -> unit ) option

Interpreter runtime errors

exception RuntimeError of run_error * ctx

val replace_undefined_with_input_variables : Mir.program -> Mir.VariableDict.t -> Mir.program

Before execution of the program, replaces the undefined stubs for input variables by their true input value

val print_output : Bir_interface.bir_function -> ctx -> unit

val raise_runtime_as_structured : run_error -> ctx -> Mir.program -> 'a

Raises a runtime error with a formatted error message and context

val compare_numbers : Mast.comp_op -> custom_float -> custom_float -> bool

Returns the comparison between two numbers in the rounding and precision context of the interpreter.

val evaluate_expr : ctx -> Mir.program -> Bir.expression Pos.marked -> value

val evaluate_program : Bir.program -> ctx -> int -> ctx