Module Mlang.Bir_to_dgfip_c

This module with a single entry point generates C files from a Bir.program.

The optimized code generation for M code, which represents the vast majority of the output, is built in DecoupledExpr.

val generate_c_program : Dgfip_options.flags -> Mir.program -> string -> unit