Module Mir.LocalVariable

Local variables don't appear in the M source program but can be introduced by let bindings when translating to MIR. They should be De Bruijn indices but instead are unique globals identifiers out of laziness.

type t = local_variable = {

id : int;

}

val new_var : unit -> t

val compare : t -> t -> int