Module Mlang.Mir_interface

While type:Mir.program contains the AST of all rules, type:full_program also encapsulates the dependency order of the rules.

val reset_all_outputs : Mir.program -> Mir.program

This function defines the var_io property of var_data to Regular for all variables that don't have Input as their var_io.

type chain_order = {

dep_graph : Mlang.Mir_dependency_graph.RG.t;

execution_order : Mir.rov_id list;

}

type full_program = {

program : Mir.program;

domains_orders : chain_order Mast.DomainIdMap.t;

chainings_orders : chain_order Mast.ChainingMap.t;

}

val to_full_program : Mir.program -> full_program

Creates the dependency graph and stores it

val output_var_dependencies : full_program -> chain_order -> Mir.variable -> unit

Print list of input variables effecting the valuation of the given variable in the given chain