Module Mir_interpreter.IntvDefInterp

type custom_float = Mir_number.IntervalNumber.t

Comes from the instantiation of the functor by a kind of floating-point value

type value =

| Number of custom_float

| Undefined

Functor-specific program values

val format_value : Stdlib.Format.formatter -> value -> unit

val format_value_prec : int -> int -> Stdlib.Format.formatter -> value -> unit

type ctx_tmp_var = {

mutable var : Com.Var.t;

mutable value : value;

}

type ctx_ref_var = {

mutable var : Com.Var.t;

mutable var_space : Com.variable_space;

mutable ref_var : Com.Var.t;

mutable org : int;

}

type print_ctx = {

mutable indent : int;

mutable is_newline : bool;

}

type ctx_var_space = {

input : value Stdlib.Array.t;

computed : value Stdlib.Array.t;

base : value Stdlib.Array.t;

}

type ctx = {

ctx_prog : Mir.program;

mutable ctx_target : Mir.target;

mutable ctx_var_space : int;

ctx_var_spaces : ctx_var_space Stdlib.Array.t;

ctx_tmps : ctx_tmp_var Stdlib.Array.t;

mutable ctx_tmps_org : int;

ctx_ref : ctx_ref_var Stdlib.Array.t;

mutable ctx_ref_org : int;

ctx_tab_map : Com.Var.t Stdlib.Array.t;

ctx_pr_out : print_ctx;

ctx_pr_err : print_ctx;

mutable ctx_anos : (Com.Error.t * string option) list;

mutable ctx_old_anos : Mlang.StrSet.t;

mutable ctx_nb_anos : int;

mutable ctx_nb_discos : int;

mutable ctx_nb_infos : int;

mutable ctx_nb_bloquantes : int;

mutable ctx_finalized_anos : (Com.Error.t * string option) list;

mutable ctx_exported_anos : (Com.Error.t * string option) list;

mutable ctx_events : (value, Com.Var.t) Com.event_value Stdlib.Array.t Stdlib.Array.t list;

}

Interpretation context

val empty_ctx : Mir.program -> ctx

val literal_to_value : Com.literal -> value

val value_to_literal : value -> Com.literal

val update_ctx_with_inputs : ctx -> Com.literal Mlang.Com.Var.Map.t -> unit

val update_ctx_with_events : ctx -> (Com.literal, Com.Var.t) Com.event_value Mlang.StrMap.t list -> unit

type run_error =

| NanOrInf of string * Mir.expression Pos.marked

| StructuredError of string * (string option * Pos.t) list * (unit -> unit) option

Interpreter runtime errors

exception RuntimeError of run_error * ctx

val raise_runtime_as_structured : run_error -> 'a

Raises a runtime error with a formatted error message and context

val compare_numbers : Com.comp_op -> custom_float -> custom_float -> bool

Returns the comparison between two numbers in the rounding and precision context of the interpreter.

val evaluate_expr : ctx -> Mir.expression Pos.marked -> value

val evaluate_program : ctx -> unit