Module Mlang.Pos

Source code position

type t

val make : string -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> t

val make_between : t -> t -> t

val format_short : Stdlib.Format.formatter -> t -> unit

val format_gnu : Stdlib.Format.formatter -> t -> unit

Respects https://www.gnu.org/prep/standards/standards.html#Formatting-Error-Messages

val format : Stdlib.Format.formatter -> t -> unit

type 'a marked =

| Mark of 'a * t
Everything related to the source code should keep its t stored, to improve error messages

val none : t

Placeholder t

val without : 'a -> 'a marked

val mark : 'a -> t -> 'a marked

val unmark : 'a marked -> 'a

val get : 'a marked -> t

val to_couple : 'a marked -> 'a * t

val map : ('a -> 'b) -> 'a marked -> 'b marked

val same : 'a -> 'b marked -> 'a marked

val unmark_option : 'a marked option -> 'a option

val get_start_line : t -> int

val get_start_column : t -> int

val get_end_line : t -> int

val get_end_column : t -> int

val get_file : t -> string

val indent_number : string -> int

val retrieve_loc_text : t -> string

Given a source code position, retrieves the content of the code by acessing the file and reading its text