Module Mlang__Bir_interface

Input-output management for BIR programs interpretation

type bir_function = {

func_variable_inputs : unit Mlang.Bir.VariableMap.t;

func_constant_inputs : Mlang.Bir.expression Mlang.Pos.marked Mlang.Bir.VariableMap.t;

func_outputs : unit Mlang.Bir.VariableMap.t;

func_conds : Mlang.Bir.condition_data Mlang.Bir.VariableMap.t;

}

Input-output data necessary to interpret a BIR program

val translate_external_conditions : Mlang.Mir.idmap -> Mlang.Mast.expression Mlang.Pos.marked list -> Mlang.Bir.condition_data Mlang.Bir.VariableMap.t

translate_external_conditions idmap conditions translates a series of boolean expressions conditions into M verification conditions ready to be added to a BIR program

val generate_function_all_vars : Mlang.Bir.program -> bir_function

Function used to generate a bir_function that includes all possible inputs and outputs

val read_function_from_spec : Mlang.Bir.program -> string -> bir_function

read_function_from_spec program spec_file reads and parses spec_file and extracts all the inputs, outputs and conditions from it.

val read_inputs_from_stdin : bir_function -> Mlang.Mir.literal Mlang.Bir.VariableMap.t

Given an input-output specification, prompts the user on stdin for the values of the inputs and returns them as a map

val adapt_program_to_function : Mlang.Bir.program -> bir_function -> Mlang.Bir.program * int

adapt_program_to_function program io modifies program according to the input-output specification of io

val context_agnostic_mpp_functions : Mlang.Bir.program -> Mlang.Bir.mpp_function Mlang.Bir.FunctionMap.t

Returns the mpp functions of the specification without contextualization from adapt_proram_to_function