Module Mlang__Bir_interpreter.IntervalInterpreter

type custom_float

Comes from the instantiation of the functor by a kind of floating-point value

type value =

| Number of custom_float

| Undefined

Functor-specific program values

val format_value : Stdlib.Format.formatter -> value -> unit

type var_value =

| SimpleVar of value

| TableVar of int * value array

Functor-specific variable values

val format_var_value : Stdlib.Format.formatter -> var_value -> unit

val format_var_value_with_var : Stdlib.Format.formatter -> (Mlang.Bir.variable * var_value) -> unit

type ctx = {

ctx_local_vars : value Mlang.Pos.marked Mlang.Mir.LocalVariableMap.t;

ctx_vars : var_value Mlang.Bir.VariableMap.t;

}

Interpretation context

val empty_ctx : ctx

val literal_to_value : Mlang.Mir.literal -> value

val var_literal_to_var_value : var_literal -> var_value

val value_to_literal : value -> Mlang.Mir.literal

val var_value_to_var_literal : var_value -> var_literal

type run_error =

| ErrorValue of string * Mlang.Pos.t

| FloatIndex of string * Mlang.Pos.t

| IndexOutOfBounds of string * Mlang.Pos.t

| IncorrectOutputVariable of string * Mlang.Pos.t

| UnknownInputVariable of string * Mlang.Pos.t

| ConditionViolated of Mlang.Mir.Error.t * Mlang.Bir.expression Mlang.Pos.marked * (Mlang.Bir.variable * var_value) list

| NanOrInf of string * Mlang.Bir.expression Mlang.Pos.marked

| StructuredError of string * (string option * Mlang.Pos.t) list * (unit -> unit) option

Interpreter runtime errors

exception RuntimeError of run_error * ctx

val replace_undefined_with_input_variables : Mlang.Mir.program -> Mlang.Mir.VariableDict.t -> Mlang.Mir.program

Before execution of the program, replaces the undefined stubs for input variables by their true input value

val raise_runtime_as_structured : run_error -> ctx -> Mlang.Mir.program -> 'a

Raises a runtime error with a formatted error message and context