Module Mlang__Bir_roundops.MainframeRoundOps

Mainframe rounding operations: those used in the mainframe context. As the behavior depends on the sie of the `long` type, this size must be given as an argument (and should be either 32 or 64).

Parameters

L : sig ... end

N : Mlang.Bir_number.NumberInterface

Signature

type t = N.t

val truncatef : t -> t

val roundf : t -> t