Module Mlang__Com

module CatVar : sig ... end

type value_typ =

| Boolean

| DateYear

| DateDayMonthYear

| DateMonth

| Integer

| Real

Here are all the types a value can have. Date types don't seem to be used at all though.

type loc_tgv = {

loc_id : string;

loc_cat : CatVar.loc;

loc_idx : int;

loc_cat_id : CatVar.t;

loc_cat_str : string;

loc_cat_idx : int;

loc_int : int;

}

type loc =

| LocTgv of string * loc_tgv

| LocTmp of string * int

| LocRef of string * int

| LocArg of string * int

| LocRes of string

module Var : sig ... end

module DomainId : Mlang.StrSet.T

module DomainIdSet : Mlang.SetSetExt.T with type base_elt = string and type T.elt = DomainId.t

module DomainIdMap : Mlang.MapExt.T with type T.key = DomainId.t

type 'a domain = {

dom_id : DomainId.t Mlang.Pos.marked;

dom_names : Mlang.Pos.t DomainIdMap.t;

dom_by_default : bool;

dom_min : DomainIdSet.t;

dom_max : DomainIdSet.t;

dom_rov : Mlang.IntSet.t;

dom_data : 'a;

dom_used : int Mlang.Pos.marked option;

}

type rule_domain_data = {

rdom_computable : bool;

}

type rule_domain = rule_domain_data domain

type verif_domain_data = {

vdom_auth : Mlang.Pos.t CatVar.Map.t;

vdom_verifiable : bool;

}

type verif_domain = verif_domain_data domain

module TargetMap : Mlang.StrMap.T

type literal =

| Float of float

| Undefined

type 'v atom =

| AtomVar of 'v

| AtomLiteral of literal

Values that can be substituted for loop parameters

type 'v set_value_loop =

| Single of 'v atom Mlang.Pos.marked

| Range of 'v atom Mlang.Pos.marked * 'v atom Mlang.Pos.marked

| Interval of 'v atom Mlang.Pos.marked * 'v atom Mlang.Pos.marked

type 'v loop_variable = char Mlang.Pos.marked * 'v set_value_loop list

A loop variable is the character that should be substituted in variable names inside the loop plus the set of value to substitute.

type 'v loop_variables =

| ValueSets of 'v loop_variable list

| Ranges of 'v loop_variable list

There are two kind of loop variables declaration, but they are semantically the same though they have different concrete syntax.

type unop =

| Not

| Minus

Unary operators

type binop =

| And

| Or

| Add

| Sub

| Mul

| Div

Binary operators

type comp_op =

| Gt

| Gte

| Lt

| Lte

| Eq

| Neq

Comparison operators

type 'v set_value =

| FloatValue of float Mlang.Pos.marked

| VarValue of 'v Mlang.Pos.marked

| Interval of int Mlang.Pos.marked * int Mlang.Pos.marked

type func =

| SumFunc
Sums the arguments

| AbsFunc
Absolute value

| MinFunc
Minimum of a list of values

| MaxFunc
Maximum of a list of values

| GtzFunc
Greater than zero (strict) ?

| GtezFunc
Greater or equal than zero ?

| NullFunc
Equal to zero ?

| ArrFunc
Round to nearest integer

| InfFunc
Truncate to integer

| PresentFunc
Different than zero ?

| Multimax
???

| Supzero
???

| VerifNumber

| ComplNumber

| Func of string

type 'v expression =

| TestInSet of bool * 'v m_expression * 'v set_value list
Test if an expression is in a set of value (or not in the set if the flag is set to false)

| Unop of unop * 'v m_expression

| Comparison of comp_op Mlang.Pos.marked * 'v m_expression * 'v m_expression

| Binop of binop Mlang.Pos.marked * 'v m_expression * 'v m_expression

| Index of 'v Mlang.Pos.marked * 'v m_expression

| Conditional of 'v m_expression * 'v m_expression * 'v m_expression option

| FuncCall of func Mlang.Pos.marked * 'v m_expression list

| FuncCallLoop of func Mlang.Pos.marked * 'v loop_variables Mlang.Pos.marked * 'v m_expression

| Literal of literal

| Var of 'v

| Loop of 'v loop_variables Mlang.Pos.marked * 'v m_expression
The loop is prefixed with the loop variables declarations

| NbCategory of Mlang.Pos.t CatVar.Map.t

| Attribut of 'v Mlang.Pos.marked * string Mlang.Pos.marked

| Size of 'v Mlang.Pos.marked

| NbAnomalies

| NbDiscordances

| NbInformatives

| NbBloquantes

and 'v m_expression = 'v expression Mlang.Pos.marked

module Error : sig ... end

type print_std =

| StdOut

| StdErr

type 'v print_arg =

| PrintString of string

| PrintName of 'v Mlang.Pos.marked

| PrintAlias of 'v Mlang.Pos.marked

| PrintIndent of 'v m_expression

| PrintExpr of 'v m_expression * int * int

type 'v formula_loop = 'v loop_variables Mlang.Pos.marked

type 'v formula_decl = 'v Mlang.Pos.marked * 'v m_expression option * 'v m_expression

type 'v formula =

| SingleFormula of 'v formula_decl

| MultipleFormulaes of 'v formula_loop * 'v formula_decl

type ('v, 'e) instruction =

| Affectation of 'v formula Mlang.Pos.marked

| IfThenElse of 'v m_expression * ('v, 'e) m_instruction list * ('v, 'e) m_instruction list

| WhenDoElse of ('v m_expression * ('v, 'e) m_instruction list * Mlang.Pos.t) list * ('v, 'e) m_instruction list Mlang.Pos.marked

| ComputeDomain of string Mlang.Pos.marked list Mlang.Pos.marked

| ComputeChaining of string Mlang.Pos.marked

| ComputeVerifs of string Mlang.Pos.marked list Mlang.Pos.marked * 'v m_expression

| ComputeTarget of string Mlang.Pos.marked * 'v Mlang.Pos.marked list

| VerifBlock of ('v, 'e) m_instruction list

| Print of print_std * 'v print_arg Mlang.Pos.marked list

| Iterate of 'v Mlang.Pos.marked * 'v Mlang.Pos.marked list * (Mlang.Pos.t CatVar.Map.t * 'v m_expression) list * ('v, 'e) m_instruction list

| Restore of 'v Mlang.Pos.marked list * ('v Mlang.Pos.marked * Mlang.Pos.t CatVar.Map.t * 'v m_expression) list * ('v, 'e) m_instruction list

| RaiseError of 'e Mlang.Pos.marked * string Mlang.Pos.marked option

| CleanErrors

| ExportErrors

| FinalizeErrors

and ('v, 'e) m_instruction = ('v, 'e) instruction Mlang.Pos.marked

val set_loc_int : loc -> int -> loc

val set_loc_tgv_cat : loc -> CatVar.loc -> string -> int -> loc

val set_loc_tgv_idx : loc -> int -> loc

val format_value_typ : Mlang.Pp.t -> value_typ -> unit

val format_literal : Mlang.Pp.t -> literal -> unit

val format_atom : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v atom -> unit

val format_loop_variables : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v loop_variables -> unit

val format_unop : Mlang.Pp.t -> unop -> unit

val format_binop : Mlang.Pp.t -> binop -> unit

val format_comp_op : Mlang.Pp.t -> comp_op -> unit

val format_set_value : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v set_value -> unit

val format_func : Mlang.Pp.t -> func -> unit

val format_expression : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v expression -> unit

val format_print_arg : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v print_arg -> unit

val format_formula : (Mlang.Pp.t -> 'v -> unit) -> Mlang.Pp.t -> 'v formula -> unit

val format_instruction : (Mlang.Pp.t -> 'v -> unit) -> (Mlang.Pp.t -> 'e -> unit) -> Mlang.Pp.t -> ('v, 'e) instruction -> unit

val format_instructions : (Mlang.Pp.t -> 'v -> unit) -> (Mlang.Pp.t -> 'e -> unit) -> Mlang.Pp.t -> ('v, 'e) m_instruction list -> unit