Module Mlang__Mir

type set_value = Mlang.Com.Var.t Mlang.Com.set_value

type access = Mlang.Com.Var.t Mlang.Com.access

type m_access = access Mlang.Pos.marked

type expression = Mlang.Com.Var.t Mlang.Com.expression

type m_expression = expression Mlang.Pos.marked

type instruction = (Mlang.Com.Var.t, Mlang.Com.Error.t) Mlang.Com.instruction

type m_instruction = instruction Mlang.Pos.marked

type target = (Mlang.Com.Var.t, Mlang.Com.Error.t) Mlang.Com.target

type stats = {

nb_computed : int;

nb_base : int;

nb_input : int;

nb_vars : int;

nb_all_tmps : int;

nb_all_refs : int;

sz_computed : int;

sz_base : int;

sz_input : int;

sz_vars : int;

sz_all_tmps : int;

nb_all_tables : int;

sz_all_tables : int;

max_nb_args : int;

table_map : Mlang.Com.Var.t Mlang.IntMap.t;

}

type program = {

program_safe_prefix : string;

program_applications : Mlang.Pos.t Mlang.StrMap.t;

program_var_categories : Mlang.Com.CatVar.data Mlang.Com.CatVar.Map.t;

program_rule_domains : Mlang.Com.rule_domain Mlang.Com.DomainIdMap.t;

program_verif_domains : Mlang.Com.verif_domain Mlang.Com.DomainIdMap.t;

program_dict : Mlang.Com.Var.t Mlang.IntMap.t;

program_vars : Mlang.Com.Var.t Mlang.StrMap.t;

program_alias : Mlang.Com.Var.t Mlang.StrMap.t;

program_var_spaces : int Mlang.StrMap.t;

program_var_spaces_idx : Mlang.Com.variable_space Mlang.IntMap.t;

program_var_space_def : Mlang.Com.variable_space;

program_event_fields : Mlang.Com.event_field Mlang.StrMap.t;

program_event_field_idxs : string Mlang.IntMap.t;

program_rules : string Mlang.IntMap.t;

program_verifs : string Mlang.IntMap.t;

program_chainings : string Mlang.StrMap.t;

program_errors : Mlang.Com.Error.t Mlang.StrMap.t;

program_functions : target Mlang.StrMap.t;

program_targets : target Mlang.StrMap.t;

program_main_target : string;

program_stats : stats;

}

val find_var_name_by_alias : program -> string Mlang.Pos.marked -> string

val find_var_by_name : program -> string Mlang.Pos.marked -> Mlang.Com.Var.t

Get a variable for a given name or alias, because of SSA multiple variables share a name or alias. If an alias is provided, the variable returned is that with the lowest execution number. When a name is provided, then the variable with the highest execution number is returned.

val expand_functions : program -> program

Calls expand_functions_expr on the whole program