Module Mlang__Mir_dependency_graph

Defines the dependency graph of an M program

module RG : Graph.Sig.G

Dependency graph for the rules of the M program. Each node corresponds to a rule, each edge to variables use. The edges in the graph go from input to outputs.

val get_used_variables_ : Mlang.Mir.expression Mlang.Pos.marked -> Mlang.Mir.VariableDict.t -> Mlang.Mir.VariableDict.t

Add all the sucessors of lvar in the graph that are used by e

val get_used_variables : Mlang.Mir.expression Mlang.Pos.marked -> Mlang.Mir.VariableDict.t

Calls the previous function with an empty dict

val create_rules_dependency_graph : Mlang.Mir.rule_data Mlang.Mir.RuleMap.t -> Mlang.Mir.rov_id Mlang.Mir.VariableMap.t -> RG.t

val check_for_cycle : RG.t -> Mlang.Mir.program -> bool -> bool

Outputs true and a warning in case of cycles.

type rule_execution_order = Mlang.Mir.rov_id list

val get_var_dependencies : ?⁠strict:bool -> Mlang.Mir.program -> rule_execution_order -> Mlang.Mir.variable -> Mlang.Mir.variable list

val pull_rules_dependencies : RG.t -> Mlang.Mir.rov_id list -> RG.t * rule_execution_order

val get_rules_execution_order : RG.t -> rule_execution_order