Module Mlang__Mir_interface

While type:Mir.program contains the AST of all rules, type:full_program also encapsulates the dependency order of the rules.

val reset_all_outputs : Mlang.Mir.program -> Mlang.Mir.program

This function defines the var_io property of var_data to Regular for all variables that don't have Input as their var_io.

type chain_order = {

dep_graph : Mlang.Mir_dependency_graph.RG.t;

execution_order : Mlang.Mir.rov_id list;

}

type full_program = {

program : Mlang.Mir.program;

chains_orders : chain_order Mlang.Mir.TagMap.t;

}

val to_full_program : Mlang.Mir.program -> Mlang.Mast.chain_tag list -> full_program

Creates the dependency graph and stores it

val output_var_dependencies : full_program -> Mlang.Mast.chain_tag -> Mlang.Mir.variable -> unit

Print list of input variables effecting the valuation of the given variable in the given chain