Module Mlang__Mir_interpreter

Interpretation of BIR programs

Program values

Instrumentation of the interpreter

val exit_on_rte : bool Stdlib.ref

If set to true, the interpreter exits the whole process in case of runtime error

val repl_debug : bool Stdlib.ref

If set to true, prints the REPL debugger in case of runtime error

The interpreter functor

module type S = sig ... end

Signature of the modules produced by the functor

module FloatDefInterp : S with type custom_float = Mlang.Mir_number.RegularFloatNumber.t

The different interpreters, which combine a representation of numbers and rounding operations. The first part of the name corresponds to the representation of numbers, and is one of the following:

module FloatMultInterp : S with type custom_float = Mlang.Mir_number.RegularFloatNumber.t

module FloatMfInterp : S with type custom_float = Mlang.Mir_number.RegularFloatNumber.t

module MPFRDefInterp : S with type custom_float = Mlang.Mir_number.MPFRNumber.t

module MPFRMultInterp : S with type custom_float = Mlang.Mir_number.MPFRNumber.t

module MPFRMfInterp : S with type custom_float = Mlang.Mir_number.MPFRNumber.t

module BigIntDefInterp : S

module BigIntMultInterp : S

module BigIntMfInterp : S

module IntvDefInterp : S with type custom_float = Mlang.Mir_number.IntervalNumber.t

module IntvMultInterp : S with type custom_float = Mlang.Mir_number.IntervalNumber.t

module IntvMfInterp : S with type custom_float = Mlang.Mir_number.IntervalNumber.t

module RatDefInterp : S with type custom_float = Mlang.Mir_number.RationalNumber.t

module RatMultInterp : S with type custom_float = Mlang.Mir_number.RationalNumber.t

module RatMfInterp : S with type custom_float = Mlang.Mir_number.RationalNumber.t

Generic interpretation API

val get_interp : Mlang.Cli.value_sort -> Mlang.Cli.round_ops -> (module S)

val evaluate_program : Mlang.Mir.program -> Mlang.Com.literal Mlang.Com.Var.Map.t -> Mlang.Cli.value_sort -> Mlang.Cli.round_ops -> Mlang.Com.literal Mlang.StrMap.t * Mlang.StrSet.t

Main interpreter function

val evaluate_expr : Mlang.Mir.program -> Mlang.Mir.expression Mlang.Pos.marked -> Mlang.Cli.value_sort -> Mlang.Cli.round_ops -> Mlang.Com.literal

Interprets only an expression