Module Mlang__Mir_typechecker

This modules defines M's static semantic. There is only one type: float. The typechecking is mostly about differentiating table from non-tables variables

val typecheck : Mlang.Mir_interface.full_program -> Mlang.Mir_interface.full_program

val expand_functions_expr : 'var Mlang.Mir.expression_ Mlang.Pos.marked -> 'var Mlang.Mir.expression_ Mlang.Pos.marked

Most functions are just syntactic sugar for operations expressible with the rest of the language, so we expand these.

val expand_functions : Mlang.Mir_interface.full_program -> Mlang.Mir_interface.full_program

Calls expand_functions_expr on the whole program