Module Utils.Cli

Command line arguments parsing

val mlang_t : (string list -> string list -> bool -> bool -> string list -> bool -> bool -> string option -> string option -> string option -> bool -> string option -> string option -> bool -> string option -> string option -> float option -> int -> bool -> string list option -> bool -> 'a) -> 'a Cmdliner.Term.t

Mlang binary command-line arguments parsing function

val info : Cmdliner.Cmd.info

Command-line man page for --help

val add_prefix_to_each_line : string -> (int -> string) -> string

add_prefix_to_each_line msg prefix will print msg but each line with line number i starts with the string prefix i

Printers

val format_with_style : ANSITerminal.style list -> ('a, unit, string) Stdlib.format -> 'a

val var_info_print : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val debug_print : ?⁠endline:string -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val warning_print : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val error_print : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val result_print : ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

val create_progress_bar : string -> (string -> unit) * (string -> unit)

Returns two functions: the first one, current_progress, has to be called during the progress loop and the other one, finish, has to be called at the end of the progressive task.