Module Mlang

module Bir_to_dgfip_c : sig ... end

This module with a single entry point generates C files from a Bir.program.

module CharMap : sig ... end

module Check_validity : sig ... end

module Cli : sig ... end

Command-line interface helpers

module Com : sig ... end

module DecoupledExpr : sig ... end

module Dgfip_compir_files : sig ... end

module Dgfip_gen_files : sig ... end

module Dgfip_m : sig ... end

module Dgfip_options : sig ... end

module Dgfip_varid : sig ... end

module Dict : sig ... end

module Driver : sig ... end

module Errors : sig ... end

module Expand_macros : sig ... end

module Format_mast : sig ... end

module Format_mir : sig ... end

module IntMap : sig ... end

module IntSet : sig ... end

module IntSetMap : sig ... end

module Irj_ast : sig ... end

module Irj_file : sig ... end

module Irj_lexer : sig ... end

module Irj_parser : sig ... end

module MapExt : sig ... end

module Mast : sig ... end

Abstract Syntax Tree for M

module Mast_to_mir : sig ... end

Mast to Mir translation of M programs.

module Mir : sig ... end

module Mir_interpreter : sig ... end

Interpretation of BIR programs

module Mir_number : sig ... end

module Mir_roundops : sig ... end

module Mlexer : sig ... end

module Mparser : sig ... end

module Parse_utils : sig ... end

Helpers for parsing

module ParserMessages : sig ... end

module Pos : sig ... end

A position in the source code is a file, as well as begin and end location of the form col:line

module Pp : sig ... end

module Prelude : sig ... end

module SetExt : sig ... end

module SetSetExt : sig ... end

module StrMap : sig ... end

module StrSet : sig ... end

module StrSetMap : sig ... end

module StrSetSet : sig ... end

module Strings : sig ... end

module Test_interpreter : sig ... end

module TopologicalSorting : sig ... end