Module Mlang.Mast

Abstract Syntax Tree

Names

type application = string

Applications are rule annotations. The 3 main DGFiP applications seem to be:

batch: deprecated, used to compute the income tax but not anymore;

bareme: seems to compute the income tax;

iliad: usage unkown, much bigger than bareme.

type chaining = string

"enchaineur" in the M source code, utility unknown

type chain_tag =

| Custom of string

| PrimCorr

| Primitif

| Corrective

| Isf

| Taux

| Irisf

| Base_hr

| Base_tl

| Base_tl_init

| Base_tl_rect

| Base_initial

| Base_inr

| Base_inr_ref

| Base_inr_tl

| Base_inr_tl22

| Base_inr_tl24

| Base_inr_ntl

| Base_inr_ntl22

| Base_inr_ntl24

| Base_inr_inter22

| Base_inr_intertl

| Base_inr_r9901

| Base_inr_cimr07

| Base_inr_cimr24

| Base_inr_cimr99

| Base_inr_tlcimr07

| Base_inr_tlcimr24

| Base_abat98

| Base_abat99

| Base_majo

| Base_premier

| Base_anterieure

| Base_anterieure_cor

| Base_stratemajo

| Non_auto_cc

| Horizontale

val all_tags : chain_tag list

val chain_tag_of_string : string -> chain_tag

val number_and_tags_of_name : string Pos.marked list -> int Pos.marked * chain_tag Pos.marked list

type variable_name = string

Variables are just strings

type func_name = string

Func names are just string for the moment

type variable_generic_name = {

base : string;

parameters : char list;

}

For generic variables, we record the list of their lowercase parameters

type error_name = string

Ununsed for now

Literals

type variable =

| Normal of variable_name

| Generic of variable_generic_name

A variable is either generic (with loop parameters) or normal

type literal =

| Variable of variable

| Float of float

type table_index =

| LiteralIndex of int

| SymbolIndex of variable

A table index is used in expressions like TABLE[X], and can be variables, integer or the special X variable that stands for a "generic" index (to define table values as a function of the index). X is contained here in SymbolIndex because there can also be a variable named "X"...

type set_value =

| FloatValue of float Pos.marked

| VarValue of variable Pos.marked

| Interval of int Pos.marked * int Pos.marked

Loops

type set_value_loop =

| Single of literal Pos.marked

| Range of literal Pos.marked * literal Pos.marked

| Interval of literal Pos.marked * literal Pos.marked

Values that can be substituted for loop parameters

type loop_variable = char Pos.marked * set_value_loop list

A loop variable is the character that should be substituted in variable names inside the loop plus the set of value to substitute.

type loop_variables =

| ValueSets of loop_variable list

| Ranges of loop_variable list

There are two kind of loop variables declaration, but they are semantically the same though they have different concrete syntax.

Expressions

type comp_op =

| Gt

| Gte

| Lt

| Lte

| Eq

| Neq

Comparison operators

type binop =

| And

| Or

| Add

| Sub

| Mul

| Div

Binary operators

val precedence : binop -> int

val has_priority : binop -> binop -> bool

val is_right_associative : 'a -> bool

val is_left_associative : binop -> bool

type unop =

| Not

| Minus

Unary operators

type expression =

| TestInSet of bool * expression Pos.marked * set_value list
Test if an expression is in a set of value (or not in the set if the flag is set to false)

| Comparison of comp_op Pos.marked * expression Pos.marked * expression Pos.marked
Compares two expressions and produce a boolean

| Binop of binop Pos.marked * expression Pos.marked * expression Pos.marked

| Unop of unop * expression Pos.marked

| Index of variable Pos.marked * table_index Pos.marked
Access a cell in a table

| Conditional of expression Pos.marked * expression Pos.marked * expression Pos.marked option
Classic conditional with an optional else clause (None only for verification conditions)

| FunctionCall of func_name Pos.marked * func_args

| Literal of literal

| Loop of loop_variables Pos.marked * expression Pos.marked
The loop is prefixed with the loop variables declarations

The main type of the M language

and func_args =

| ArgList of expression Pos.marked list

| LoopList of loop_variables Pos.marked * expression Pos.marked

Functions can take a explicit list of argument or a loop expression that expands into a list

Toplevel clauses

Rules

type lvalue = {

var : variable Pos.marked;

index : table_index Pos.marked option;

}

An lvalue (left value) is a variable being assigned. It can be a table or a non-table variable

type formula_decl = {

lvalue : lvalue Pos.marked;

formula : expression Pos.marked;

}

type formula =

| SingleFormula of formula_decl

| MultipleFormulaes of loop_variables Pos.marked * formula_decl

In the M language, you can define multiple variables at once. This is the way they do looping since the definition can depend on the loop variable value (e.g Xi can depend on i).

type rule = {

rule_number : int Pos.marked;

rule_tags : chain_tag Pos.marked list;

rule_applications : application Pos.marked list;

rule_chaining : chaining Pos.marked option;

rule_formulaes : formula Pos.marked list;
A rule can contain many variable definitions

}

Variable declaration

Input variables

type input_variable_subtype =

| Context

| Family

| Penality

| Income

Unused for now, except for typechecking: Income should be a real number corresponding to an amount of money

type input_variable_attribute = string

Attributes are unused for now

type value_typ =

| Boolean

| DateYear

| DateDayMonthYear

| DateMonth

| Integer

| Real

Here are all the types a value can have. Date types don't seem to be used at all though.

type input_variable = {

input_name : variable_name Pos.marked;

input_subtyp : input_variable_subtype Pos.marked;

input_attributes : (input_variable_attribute Pos.marked * literal Pos.marked) list;

input_given_back : bool;
An input variable given back ("restituee") means that it's also an output

input_alias : variable_name Pos.marked;
Unused for now

input_description : string Pos.marked;

input_typ : value_typ Pos.marked option;

}

type computed_typ =

| Base

| GivenBack

A GivenBack variable is an output of the program

type computed_variable = {

comp_name : variable_name Pos.marked;

comp_table : int Pos.marked option;
size of the table, None for non-table variables

comp_attributes : (input_variable_attribute Pos.marked * literal Pos.marked) list;

comp_subtyp : computed_typ Pos.marked list;

comp_typ : value_typ Pos.marked option;

comp_description : string Pos.marked;

}

type variable_decl =

| ComputedVar of computed_variable Pos.marked

| ConstVar of variable_name Pos.marked * literal Pos.marked
The literal is the constant value

| InputVar of input_variable Pos.marked

Verification clauses

type verification_condition = {

verif_cond_expr : expression Pos.marked;

verif_cond_error : error_name Pos.marked * variable_name Pos.marked option;
A verification condition error can ba associated to a variable

}

type verification = {

verif_number : int Pos.marked;

verif_tags : chain_tag Pos.marked list;

verif_applications : application Pos.marked list;
Verification conditions are application-specific

verif_conditions : verification_condition Pos.marked list;

}

type error_typ =

| Anomaly

| Discordance

| Information

val compare_error_type : error_typ -> error_typ -> int

type error_ = {

error_name : error_name Pos.marked;

error_typ : error_typ Pos.marked;

error_descr : string Pos.marked list;

}

M programs

type source_file_item =

| Application of application Pos.marked
Declares an application

| Chaining of chaining * application Pos.marked list
Unused, declares an "enchaineur"

| VariableDecl of variable_decl

| Rule of rule

| Verification of verification

| Error of error_
Declares an error

| Output of variable_name Pos.marked
Declares an output variable

| Function
Declares a function, unused

type source_file = source_file_item Pos.marked list

type program = source_file list

Function specification AST

type function_spec = {

spec_inputs : variable_name Pos.marked list;

spec_consts : (variable_name Pos.marked * expression Pos.marked) list;

spec_outputs : variable_name Pos.marked list;

spec_conditions : expression Pos.marked list;

}

Helper functions

val get_variable_name : variable -> string

val are_tags_part_of_chain : chain_tag list -> chain_tag -> bool

val are_tags_part_of_verif_chain : chain_tag list -> chain_tag -> bool