Module Mir_interpreter.IntvDefInterp

type custom_float = Mir_number.IntervalNumber.t

Comes from the instantiation of the functor by a kind of floating-point value

type value =

| Number of custom_float

| Undefined

Functor-specific program values

val format_value : Stdlib.Format.formatter -> value -> unit

val format_value_prec : int -> int -> Stdlib.Format.formatter -> value -> unit

type print_ctx = {

mutable indent : int;

mutable is_newline : bool;

}

type ctx = {

ctx_tgv : value Stdlib.Array.t;

ctx_tmps : value Stdlib.Array.t;

mutable ctx_tmps_org : int;

ctx_ref : (Com.Var.t * int) Stdlib.Array.t;

mutable ctx_ref_org : int;

mutable ctx_args : value Stdlib.Array.t list;

mutable ctx_res : value list;

ctx_pr_out : print_ctx;

ctx_pr_err : print_ctx;

mutable ctx_anos : (Com.Error.t * string option) list;

mutable ctx_old_anos : StrSet.t;

mutable ctx_nb_anos : int;

mutable ctx_nb_discos : int;

mutable ctx_nb_infos : int;

mutable ctx_nb_bloquantes : int;

mutable ctx_finalized_anos : (Com.Error.t * string option) list;

mutable ctx_exported_anos : (Com.Error.t * string option) list;

}

Interpretation context

val empty_ctx : Mir.program -> ctx

val literal_to_value : Com.literal -> value

val value_to_literal : value -> Com.literal

val update_ctx_with_inputs : ctx -> Com.literal Com.Var.Map.t -> unit

type run_error =

| NanOrInf of string * Mir.expression Pos.marked

| StructuredError of string * (string option * Pos.t) list * ( unit -> unit ) option

Interpreter runtime errors

exception RuntimeError of run_error * ctx

val raise_runtime_as_structured : run_error -> 'a

Raises a runtime error with a formatted error message and context

val compare_numbers : Com.comp_op -> custom_float -> custom_float -> bool

Returns the comparison between two numbers in the rounding and precision context of the interpreter.

val evaluate_expr : ctx -> Mir.program -> Mir.expression Pos.marked -> value

val evaluate_program : Mir.program -> ctx -> unit