Module type Mir_roundops.RoundOpsFunctor

The actual implementation of rounding operations depends on the chosen representation of numbers, hence we need a functor

Parameters

module N : Mir_number.NumberInterface

Signature

type t = N.t

val truncatef : t -> t

val roundf : t -> t