Module Mlang.Strings

val sanitize_str : (string * Pos.t) -> string

DGFiP sources are encoded in iso-8859-1 which is not compatible with some backend compilers such as Java and Python, this function transforms illegal characters with a space.