Module Mlang__Bir_interpreter

Interpretation of BIR programs

Program values

type var_literal =

| SimpleVar of Mlang.Mir.literal

| TableVar of int * Mlang.Mir.literal array

Instrumentation of he interpreter

type code_location_segment =

| InsideBlock of int

| ConditionalBranch of bool

| InsideRule of Mlang.Bir.rov_id

| InsideFunction of Mlang.Bir.function_name

Representation of each program location segment

val format_code_location_segment : Stdlib.Format.formatter -> code_location_segment -> unit

type code_location = code_location_segment list

A program location is simply the path inside the program

val format_code_location : Stdlib.Format.formatter -> code_location -> unit

val assign_hook : (Mlang.Bir.variable -> (unit -> var_literal) -> code_location -> unit) Stdlib.ref

The instrumentation of the interpreter is done through this reference. The function that you assign to this reference will be called each time a variable assignment is executed

val exit_on_rte : bool Stdlib.ref

If set to true, the interpreter exits the whole process in case of runtime error

val repl_debug : bool Stdlib.ref

If set to true, prints the REPL debugger in case of runtime error

The interpreter functor

module type S = sig ... end

Signature of the modules produced by the functor

module RegularFloatInterpreter : S

module MPFRInterpreter : S

module BigIntInterpreter : S

module IntervalInterpreter : S

module RationalInterpreter : S

Generic interpretation API

type value_sort =

| RegularFloat

| MPFR of int
bitsize of the floats

| BigInt of int
precision of the fixed point

| Interval

| Rational

According on the value_sort, a specific interpreter will be called with the right kind of floating-point value

val evaluate_program : Mlang.Bir_interface.bir_function -> Mlang.Bir.program -> Mlang.Mir.literal Mlang.Bir.VariableMap.t -> int -> value_sort -> unit -> unit

Main interpreter function

val evaluate_expr : Mlang.Mir.program -> Mlang.Bir.expression Mlang.Pos.marked -> value_sort -> Mlang.Mir.literal

Interprets only an expression