Module Mlang__Mast_to_mir

Mast to Mir translation of M programs.

Translation context

Loop translation context

type loop_param_value =

| VarName of Mlang.Mast.variable_name

| RangeInt of int

The values of the map can be either strings of integers

module ConstMap : Stdlib.Map.S with type ConstMap.key = Mlang.Mast.variable_name

module ParamsMap : Stdlib.Map.S with type ParamsMap.key = Stdlib.Char.t

Map whose keys are loop parameters

type loop_context = (loop_param_value * int) ParamsMap.t

This is the context when iterating a loop : for each loop parameter, we have access to the current value of this loop parameter in this iteration.

General translation context

type translating_context = {

table_definition : bool;
true if translating an expression susceptible to contain a generic table index

idmap : Mlang.Mir.idmap;
Current string-to-Mir.Variable.t mapping

lc : loop_context option;
Current loop translation context

const_map : float Mlang.Pos.marked ConstMap.t;
Mapping from constant variables to their value

exec_number : Mlang.Mir.execution_number;
Number of the rule of verification condition being translated

}

Translation helpers

val get_var_from_name : Mlang.Mir.idmap -> string Mlang.Pos.marked -> Mlang.Mir.execution_number -> bool -> Mlang.Mir.Variable.t

Queries a type: Mir.variable.t from an type:idmap mapping, the name of a variable and the rule number from which the variable is requested. Returns the variable with the same name and highest rule number that is below the current rule number from where this variable is requested

val list_max_execution_number : Mlang.Mir.Variable.t list -> Mlang.Mir.Variable.t

Helper to compute the max SSA candidate in a list

val translate_expression : translating_context -> Mlang.Mast.expression Mlang.Pos.marked -> Mlang.Mir.expression Mlang.Pos.marked

Main translation function for expressions

val dummy_exec_number : Mlang.Pos.t -> Mlang.Mir.execution_number

Dummy execution number used for variable declarations

val get_conds : Mlang.Mir.Error.t list -> float Mlang.Pos.marked ConstMap.t -> Mlang.Mir.idmap -> Mlang.Mast.program -> Mlang.Mir.condition_data Mlang.Mir.VariableMap.t

Returns a map whose keys are dummy variables and whose values are the verification conditions.

Main translation function

val translate : Mlang.Mast.program -> Mlang.Mir.program

Main translation function from the M AST to the M Variable Graph. This function performs 6 linear passes on the input code:

get_constants gets the value of all constant variables, the values of which are needed to compute certain loop bounds;

get_variables_decl retrieves the declarations of all other variables and errors;

get_var_redefinitions incorporates into idmap all definitions inside rules along with their execution number;

get_var_data is the workhorse pass that translates all the expressions corresponding to the definitions;

add_dummy_definition_for_variable_declaration adds Undefined definitions placeholder for all variables declarations;

get_errors_conds parses the verification conditions definitions.